requirements.txt
asgiref==3.8.1
black==24.3.0
certifi==2024.7.4
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==44.0.1
defusedxml==0.7.1
Django==5.2
django-cors-headers==3.8.0
django-filter==21.1
django-phonenumber-field==6.0.0
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
djoser==2.3.1
drf-nested-routers==0.93.4
drf-yasg==1.21.10
idna==3.7
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
mypy-extensions==0.4.3
oauthlib==3.2.2
packaging==25.0
pathspec==0.9.0
phonenumbers==8.12.41
pillow==11.2.1
platformdirs==2.5.1
psycopg2-binary==2.9.10
pycparser==2.21
PyJWT==2.10.1
pyparsing==3.0.7
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0.2
requests==2.32.0
requests-oauthlib==1.3.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.6.1
sqlparse==0.5.0
tomli==2.0.1
uritemplate==4.1.1
urllib3==2.2.2
